requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
Werkzeug
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
python-dotenv==1.0.1
